MONEYMONEY

+++ formated.py
@@ -31,7 +31,7 @@ def formatExperts(listexp): #final list of experts to itenerate on
     for i in listexp:
         i[2] = tuple(i[2].replace(";",",").replace("(","").replace(")","").split(","))
         i[3] = int(i[3])
-        i[4] = float(i[4])
+        i[4] = int(i[4])
         i[7] = float(i[7])
         
     return listexp
@@ -43,7 +43,7 @@ def formatClients(listcl): #final list of clients to itenerate on
     
     listcl = takeSpaces(listcl)        
     for i in listcl:
-        i[4] = float(i[4])
+        i[4] = int(i[4])
         i[5] = int(i[5])
         i[7] = int(i[7][0])*60 + int(i[7][2:4])  #horas em minutos
